day1_tip_calculator.py
Just finished module 2 in the NetAcad Python Essentials 1 (PE1) and decided to use my skills to create this mini-project that calculates the tip amount in accordance with the bill amount. I am not sure right now but it the future I will come back to this mini-project to polish the code as I know that it isn't very efficient and clean but it gets the job done.