Add files via upload
I tried to create a password checker where a score is calculated whenever conditions are met or not. My current limited knowledge didn't really allow me to do this right now, but after I finish the PE1 course I think that I would be able to give it another go and make it functional.