day3_secret_number
This is a LAB section of the PE1 Course where there was a given scenario and I had to complete the code for it. I think think that this is a successful program because it is functional, quite economical in regards to the code usage and also it was a really good practice on while loops.